requirements.txt
# Flask and Server
Flask==3.0.0
gunicorn==21.2.0

# Data Processing
pandas==2.1.4

# Database
mysql-connector-python==8.2.0

# Utilities
python-dotenv==1.0.0

